Backend/requirements.txt
setuptools
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
authlib
itsdangerous
motor==3.6.0
pydantic[email]
asyncio
httpx
fastapi[standard]
websockets
slowapi
redis==4.5.5
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi-sso
asgiref
pymongo
fastapi-cache2[redis]
jinja2
pyjwt